refactor: split up code into separate modules to make it easier to change
