+++ src/parts/ErrorCodes/ErrorCodes.js
@@ -1,0 +1,2 @@
+export const ENOENT = 'ENOENT'
+export const ERR_MODULE_NOT_FOUND = 'ERR_MODULE_NOT_FOUND'
