+++ src/parts/InvariantError/InvariantError.js
@@ -1,0 +1,5 @@
+export class InvariantError extends Error {
+  constructor(message) {
+    super(message)
+  }
+}
