+++ src/parts/Logger/Logger.js
@@ -1,0 +1,7 @@
+export const info = (...args) => {
+  console.info(...args)
+}
+
+export const error = (...args) => {
+  console.error(...args)
+}
